Replace JSX for buttons in SideNav with SideNavButton component instances.
