--- src/layout/SideNav/SideNav-styled.tsx
@@ -1,25 +1,0 @@
-import { Box, IconButton, Typography } from "@mui/material";
-import { styled } from "@mui/material/styles";
-
-export const StyledIconButton = styled(IconButton)(({ theme }) => ({
-  color: "#F4EDED",
-  display: "flex",
-  flexDirection: "column",
-  gap: theme.spacing(2),
-  cursor: "pointer",
-  "&:hover": {
-    color: "#E36414",
-    outline: "none",
-  }
-}));
-
-export const StyledTypography = styled(Typography)(({ theme }) => ({
-  fontSize: theme.spacing(1.65),
-  color: "#F4EDED",
-}));
-
-export const StyledButtonContainer = styled(Box)(({
-  display: "flex",
-  flexDirection: "column",
-  margin: "0 auto",
-}));
